Made almost all external-local functions 'static'
so they don't pollute the global namespace.
we MUST export the _setup() functions, so Pd can find them when loading the library.
all other functions (that are not referenced in other files) should be declared 'static',
so they are not exported at all.

also: removed (static) function declarations that missed a function *definition*.