updated "fftinfo" message
Removed references to 32-bit, and MSP vector. Also removed library version, since the library is contained in this repo, so will always be in sync. The "fftinfo" message reports incorrect values until Pd dsp is turned on. It would be great to improve initialization, such that the signal vector size, and FFT window size are correctly initialized prior to turning on the DACs.