updated buffer copy and zero pad
New implementation of copying & zero-padding FFT data into a temporary buffer that allows the spectrum to stretch or shrink. This is done in the do_mindwarp() function. Removed an error message on illegal warp factors, for cosmetic purposes. (The help file already specifies the legal limits, so diagnostic messages are not strictly needed here.) Fixed a crashable bug where the warp factor is set between 0 - 1/16 (legal lower limit). Now, illegal warp factors are reset to 1.